feat: MiddlewareManager TS 마이그레이션
