+++ src/utils/Routes.js
@@ -1,7 +1,7 @@
 import {URL} from 'url';
 import {AppError} from "./AppError.ts";
 import {ResponseHandler} from "./ResponseHandler.js";
-import {MiddlewareManager} from "../middlewares/MiddlewareManager.js";
+import {MiddlewareManager} from "../middlewares/MiddlewareManager.ts";
 
 export class Router {
     constructor() {
