+++ tsconfig.json
@@ -2,17 +2,20 @@
   "compilerOptions": {
     "target": "ESNext",
     "module": "ESNext",
-    "outDir": "dist",
+    "outDir": "./dist",
+    "moduleResolution": "Node",
+    "allowSyntheticDefaultImports": true,
+    "allowImportingTsExtensions": false,
+    "resolveJsonModule": true,
+
     "moduleDetection": "force",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true
   },
-  "ts-node": {
-    "esm": true
-  },
   "include": [
-    "src"
+    "src",
+    "**/*.ts"
   ]
 }
